feat(wasm): add TypeScript type definitions and reorganize WASM bindings
- Added a comprehensive `typescript_custom_section` in `src/wasm.rs` defining interfaces for better TypeScript integration.
- Reorganized `src/wasm.rs` by moving several methods to `src/wasm_ext.rs` to maintain the 500 LOC limit.
- Added `typescript_type` annotations to WASM-exposed methods to improve the generated `.d.ts` file.
- Updated visibility of shared types and constants to `pub(crate)`.
